refactor: 도메인에 따른 Controller 분리
